move estimate_stabilization
